update timeout and model
