+++ back-end/go.mod
@@ -2,9 +2,11 @@ module lcs_backend
 
 go 1.18
 
+require github.com/gin-contrib/cors v1.3.1
+
 require (
-	github.com/gin-contrib/cors v1.3.1
 	github.com/gin-contrib/sse v0.1.0 // indirect
+	github.com/gin-contrib/timeout v0.0.3 // indirect
 	github.com/gin-gonic/gin v1.8.1 // indirect
 	github.com/go-playground/locales v0.14.0 // indirect
 	github.com/go-playground/universal-translator v0.18.0 // indirect
