+++ back-end/model/user.go
@@ -11,3 +11,26 @@ type User struct {
 	enabled bool
 
 }
+type Password struct {
+	Password string 
+}
+type UserUpdatePassword struct {
+	CurrentPassword string
+	NewPassword string
+}
+type Number struct {
+	Number int
+}
+type UserRead struct {
+	Name string `json:"name"`
+	Email	 string  `json:"email"`
+}
+type UserCreate struct {
+	Email	 string  
+	Name 	 string  
+	Password string  
+}
+type UserUpdate struct {
+	Name string
+
+}
