+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/YOUR-USER-OR-ORG-NAME/YOUR-REPO-NAME
 
 go 1.14
+
+require github.com/gin-contrib/timeout v0.0.3 // indirect
